requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=2.0.0
pyyaml>=6.0
python-dotenv>=0.19.0